third_party/blink/web_tests/fast/forms/selection-start-end-readonly-expected.txt

Tests for selectionStart and selectionEnd on read-only INPUT and TEXTAREA.
bug 25444: SelectionStart, selectionEnd properties return wrong values when the selection is in a form input.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS element.selectionStart is 0
PASS element.selectionEnd is 10
PASS element.selectionStart is 0
PASS element.selectionEnd is 9
PASS element.selectionStart is 1
PASS element.selectionEnd is 10
PASS element.selectionStart is 5
PASS element.selectionEnd is 5
PASS element.selectionStart is 5
PASS element.selectionEnd is 5
PASS element.selectionStart is 0
PASS element.selectionEnd is 0
PASS element.selectionStart is 10
PASS element.selectionEnd is 10
PASS element.selectionStart is 0
PASS element.selectionEnd is 10
PASS element.selectionStart is 0
PASS element.selectionEnd is 9
PASS element.selectionStart is 1
PASS element.selectionEnd is 10
PASS element.selectionStart is 5
PASS element.selectionEnd is 5
PASS element.selectionStart is 5
PASS element.selectionEnd is 5
PASS element.selectionStart is 0
PASS element.selectionEnd is 0
PASS element.selectionStart is 10
PASS element.selectionEnd is 10
PASS successfullyParsed is true

TEST COMPLETE
